Create Game#validate_selected_square
This method takes one argument, a player instance object, and uses Player#select piece to prompt the player to select a square to pick one of their pieces from.

Tests have been written for this method to ensure it works correctly when player enters in a square that does not exist.